requirements.txt
PyGithub>=2.1.1
PyYAML>=6.0.1
crewai>=0.1.0
pydantic>=2.5.0
